+++ superwall/src/main/java/com/superwall/sdk/config/options/PaywallOptions.kt
@@ -1,5 +1,7 @@
 package com.superwall.sdk.config.options
 
+import kotlin.time.Duration
+
 // Options for configuring the appearance and behavior of paywalls.
 class PaywallOptions {
     // Determines whether the paywall should use haptic feedback. Defaults to true.
@@ -86,6 +88,10 @@ class PaywallOptions {
 
     // Hide shimmer optimistically
     var optimisticLoading: Boolean = false
+
+    // How long until a paywall timeout is invoked.
+    // If not using fallback loading, settings this will trigger a paywall timeout instead of retrying.
+    var timeoutAfter: Duration? = null
 }
 
 internal fun PaywallOptions.toMap() =