+++ superwall/src/main/java/com/superwall/sdk/paywall/presentation/internal/state/PaywallErrors.kt
@@ -1,0 +1,9 @@
+package com.superwall.sdk.paywall.presentation.internal.state
+
+sealed class PaywallErrors(
+    val msg: String?,
+) : Throwable(msg) {
+    class Timeout(
+        message: String?,
+    ) : PaywallErrors(message)
+}